Some big changes in solution explorer in order to keep it 'tidy'
