+++ SuppLocals/SuppLocals/Classes/EmailSender.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using SuppLocals.Views;
+using System;
 using System.Collections.Generic;
 using System.Net;
 using System.Net.Mail;
