+++ SuppLocals/SuppLocals/Classes/Enums/VendorType.cs
