+++ SuppLocals/SuppLocals/Classes/ObservableObject.cs
