+++ SuppLocals/SuppLocals/Classes/PasswordHelper.cs
@@ -1,10 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Windows;
+﻿using System.Windows;
 using System.Windows.Controls;
 
-namespace SuppLocals.Helper
+namespace SuppLocals
 {
     public static class PasswordHelper
     {
