--- SuppLocals/SuppLocals/TabsModel/AboutModel.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace SuppLocals.TabsModel
-{
-    class AboutModel
-    {
-    }
-}
