+++ SuppLocals/SuppLocals/ViewModels/AboutVM.cs
@@ -1,0 +1,7 @@
+﻿
+namespace SuppLocals.ViewModels
+{
+    public class AboutVM
+    {
+    }
+}
