+++ SuppLocals/SuppLocals/ViewModels/AddServiceVM.cs
@@ -1,0 +1,8 @@
+﻿
+
+namespace SuppLocals.ViewModels
+{
+    public class AddServiceVM
+    {
+    }
+}
