+++ SuppLocals/SuppLocals/ViewModels/FaqVM.cs
@@ -1,0 +1,7 @@
+﻿
+namespace SuppLocals.ViewModels
+{
+    public class FaqVM
+    {
+    }
+}
