+++ SuppLocals/SuppLocals/ViewModels/MyServicesVM.cs
@@ -1,0 +1,6 @@
+﻿namespace SuppLocals.ViewModels
+{
+    public class MyServicesVM
+    {
+    }
+}
