+++ SuppLocals/SuppLocals/ViewModels/VendorsVM.cs
@@ -1,0 +1,7 @@
+﻿
+namespace SuppLocals.ViewModels
+{
+    public class VendorsVM
+    {
+    }
+}
