+++ SuppLocals/SuppLocals/Views/AccountViews/Login.xaml
@@ -1,10 +1,9 @@
-﻿<Window x:Class="SuppLocals.Login"
+﻿<Window x:Class="SuppLocals.Views.Login"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:SuppLocals"
-        xmlns:helper="clr-namespace:SuppLocals.Helper"
         mc:Ignorable="d"
         Title="Login" Height="500" Width="400"
         AllowsTransparency="true" Background="Transparent"
@@ -76,8 +75,8 @@
                         FontWeight="Bold"
                         Margin="70 0 0 0 "/>
 
-                <PasswordBox  helper:PasswordHelper.Attach="True"
-                        helper:PasswordHelper.Password="{Binding Password,
+                <PasswordBox  local:PasswordHelper.Attach="True"
+                        local:PasswordHelper.Password="{Binding Password,
                         Mode=TwoWay,  UpdateSourceTrigger=PropertyChanged, 
                         ValidatesOnDataErrors=True}"
                         ToolTip="{Binding ErrorCollection[Password]}"
