+++ SuppLocals/SuppLocals/Views/AccountViews/SignUp.xaml
@@ -1,11 +1,9 @@
-﻿<Window x:Class="SuppLocals.SignUp"
+﻿<Window x:Class="SuppLocals.Views.SignUp"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:SuppLocals"
-        xmlns:helper="clr-namespace:SuppLocals.Helper"
-        
+        xmlns:local="clr-namespace:SuppLocals"        
         mc:Ignorable="d"
         Title="SignUp" Height="500" Width="400"
         AllowsTransparency="true" Background="Transparent"
@@ -122,8 +120,8 @@
                         FontWeight="Bold"
                         Margin="60 0 0 0 "/>
 
-                <PasswordBox helper:PasswordHelper.Attach="True"
-                        helper:PasswordHelper.Password="{Binding Password,
+                <PasswordBox local:PasswordHelper.Attach="True"
+                        local:PasswordHelper.Password="{Binding Password,
                         Mode=TwoWay,  UpdateSourceTrigger=PropertyChanged, 
                         ValidatesOnDataErrors=True}"
                         ToolTip="{Binding ErrorCollection[Password]}"
@@ -159,8 +157,8 @@
                         FontWeight="Bold"
                         Margin="60 0 0 0 "/>
 
-                <PasswordBox  helper:PasswordHelper.Attach="True"
-                        helper:PasswordHelper.Password="{Binding ConfirmPassword,
+                <PasswordBox  local:PasswordHelper.Attach="True"
+                        local:PasswordHelper.Password="{Binding ConfirmPassword,
                         Mode=TwoWay,  UpdateSourceTrigger=PropertyChanged, 
                         ValidatesOnDataErrors=True}"
                         ToolTip="{Binding ErrorCollection[ConfirmPassword]}"
