+++ SuppLocals/SuppLocals/Views/TabsViews/About.xaml
@@ -1,9 +1,9 @@
-﻿<UserControl x:Class="SuppLocals.Tabs.About"
+﻿<UserControl x:Class="SuppLocals.Views.About"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:SuppLocals.Tabs"
+             xmlns:local="clr-namespace:SuppLocals"
              xmlns:images="SuppLocals.Assets"
              mc:Ignorable="d" 
              d:DesignHeight="340" d:DesignWidth="800" >
