+++ SuppLocals/SuppLocals/Views/TabsViews/About.xaml.cs
@@ -1,0 +1,14 @@
+﻿
+using System.Windows.Controls;
+
+namespace SuppLocals.Views
+{
+
+    public partial class About : UserControl
+    {
+        public About()
+        {
+            InitializeComponent();
+        }
+    }
+}
