+++ SuppLocals/SuppLocals/Views/TabsViews/Home.xaml
@@ -1,25 +1,28 @@
-﻿<UserControl x:Class="SuppLocals.Tabs.Vendors"
+﻿<UserControl x:Class="SuppLocals.Views.Home"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:SuppLocals.Tabs"
+             xmlns:m="clr-namespace:Microsoft.Maps.MapControl.WPF;assembly=Microsoft.Maps.MapControl.WPF"
+             xmlns:local="clr-namespace:SuppLocals"
              mc:Ignorable="d" 
-   d:DesignHeight="340" d:DesignWidth="800">
+    d:DesignHeight="340" d:DesignWidth="800">
+   
     <Grid Background="#CCBA8B">
         <Grid.ColumnDefinitions>
             <ColumnDefinition Width="*"/>
-            <ColumnDefinition Width="auto"/>
+            <ColumnDefinition Width="*"/>
             <ColumnDefinition Width="*"/>
         </Grid.ColumnDefinitions>
         <Grid.RowDefinitions>
-            <RowDefinition Height="auto"/>
             <RowDefinition Height="*"/>
-            <RowDefinition Height="40"/>
         </Grid.RowDefinitions>
-
-        <TextBlock Grid.Row="0" Grid.Column="1" Text="DISCOVER ALL VENDORS" VerticalAlignment="Top" HorizontalAlignment="Center" FontSize="20" FontWeight="Bold" Margin="10 10 10 10"/>
         
+        <!--Map-->
+        <m:Map Grid.Column="0" Grid.ColumnSpan="3" Grid.Row="0"  x:Name="myMap" Mode="Road" Center="54.68,25.27" ZoomLevel="12">
+            <m:MapLayer x:Name="RouteLayer" HorizontalAlignment="Right" Width="745"/>
+        </m:Map>
+
 
     </Grid>
 </UserControl>
