+++ SuppLocals/SuppLocals/Views/TabsViews/MyServices.xaml
@@ -1,9 +1,9 @@
-﻿<UserControl x:Class="SuppLocals.Tabs.MyServices"
+﻿<UserControl x:Class="SuppLocals.Views.MyServices"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:SuppLocals.Tabs"
+             xmlns:local="clr-namespace:SuppLocals"
              mc:Ignorable="d" 
              d:DesignHeight="340" d:DesignWidth="800">
     <Grid Background="Purple">
