+++ SuppLocals/SuppLocals/Views/TabsViews/MyServices.xaml.cs
@@ -1,0 +1,16 @@
+﻿using System.Windows.Controls;
+
+
+namespace SuppLocals.Views
+{
+    /// <summary>
+    /// Interaction logic for MyServices.xaml
+    /// </summary>
+    public partial class MyServices : UserControl
+    {
+        public MyServices()
+        {
+            InitializeComponent();
+        }
+    }
+}
