+++ SuppLocals/SuppLocals/Views/TabsViews/Vendors.xaml.cs
@@ -1,0 +1,16 @@
+﻿using System.Windows.Controls;
+
+
+namespace SuppLocals.Views
+{
+    /// <summary>
+    /// Interaction logic for Vendors.xaml
+    /// </summary>
+    public partial class Vendors : UserControl
+    {
+        public Vendors()
+        {
+            InitializeComponent();
+        }
+    }
+}
